fixed issue with migration and sql column type

+++ cmd/migrate/migrations/20260423095328_add_version_posts.sql
@@ -1,5 +1,5 @@
 -- +goose Up
-ALTER TABLE posts ADD COLUMN version INTEGER NOT NULL DEFAULT 0;
+ALTER TABLE posts ADD COLUMN version BIGINT NOT NULL DEFAULT 0;
 
 -- +goose Down
 ALTER TABLE posts DROP COLUMN version;
